# Conflicts:
#	glm/detail/func_geometric.inl
#	glm/detail/func_trigonometric.inl
#	glm/detail/setup.hpp
#	glm/detail/type_mat2x2.hpp
#	glm/detail/type_mat2x2.inl
#	glm/detail/type_mat2x3.hpp
#	glm/detail/type_mat2x3.inl
#	glm/detail/type_mat2x4.hpp
#	glm/detail/type_mat2x4.inl
#	glm/detail/type_mat3x2.hpp
#	glm/detail/type_mat3x2.inl
#	glm/detail/type_mat3x3.hpp
#	glm/detail/type_mat3x3.inl
#	glm/detail/type_mat3x4.hpp
#	glm/detail/type_mat3x4.inl
#	glm/detail/type_mat4x2.hpp
#	glm/detail/type_mat4x2.inl
#	glm/detail/type_mat4x3.hpp
#	glm/detail/type_mat4x3.inl
#	glm/detail/type_mat4x4.hpp
#	glm/detail/type_mat4x4.inl
#	glm/detail/type_vec4.hpp
#	glm/detail/type_vec4.inl
#	glm/ext/matrix_relational.inl
#	glm/geometric.hpp
#	glm/gtc/integer.hpp
#	glm/gtc/integer.inl
#	glm/gtc/reciprocal.hpp
#	glm/gtc/reciprocal.inl
#	glm/gtc/type_ptr.inl
#	glm/gtx/associated_min_max.inl
#	glm/gtx/extended_min_max.inl
#	glm/gtx/wrap.inl

+++ glm/gtx/scalar_multiplication.hpp
@@ -1,6 +1,9 @@
-/// @ref gtx
+/// @ref gtx_scalar_multiplication
 /// @file glm/gtx/scalar_multiplication.hpp
 /// @author Joshua Moerman
+///
+/// @defgroup gtx_scalar_multiplication GLM_GTX_scalar_multiplication
+/// @ingroup gtx
 ///
 /// Include <glm/gtx/scalar_multiplication.hpp> to use the features of this extension.
 ///
@@ -32,6 +35,9 @@
 
 namespace glm
 {
+	/// @addtogroup gtx_scalar_multiplication
+	/// @{
+
 	template<typename T, typename Vec>
 	using return_type_scalar_multiplication = typename std::enable_if<
 		!std::is_same<T, float>::value       // T may not be a float
@@ -54,7 +60,7 @@ namespace glm
 	template<typename T> \
 	return_type_scalar_multiplication<T, Vec> \
 	operator/(Vec lh, T const& s){ \
-		return lh *= 1.0f / s; \
+		return lh *= 1.0f / static_cast<float>(s); \
 	}
 
 GLM_IMPLEMENT_SCAL_MULT(vec2)
@@ -72,4 +78,5 @@ GLM_IMPLEMENT_SCAL_MULT(mat4x3)
 GLM_IMPLEMENT_SCAL_MULT(mat4)
 
 #undef GLM_IMPLEMENT_SCAL_MULT
+	/// @}
 } // namespace glm
